Uploaded local copy of Keras model
Some users will not have access to the AI model used for this
application, whether it is due to lack of computational resources
or (have to be honest) laziness. I have imported the model in this
repo to alleviate these situations and allow the user to run
the program as expected without any struggle over webapp usage
(e.g., no predictions coming out & erroring webpage).