Fixed critical error with model not predicting
The provided model failed to make a prediction properly due to
a missing anxiety factor that I did not update the model to
include previously. The new model should be able to predict
the GAD-7 score without any issues. Not sure if there are any
other errors besides this, but I will be on the lookout if the
webapp ceases to function after this commit.

Additionally, I have added some minor edits for coding convenience,
should you wish to look at the source code!